+++ Common/Debug/Common.log
@@ -1,6 +1,1 @@
-﻿  MatrixStack.cpp
-  Mtx44.cpp
-  timer.cpp
-  Vector3.cpp
-  Generating Code...
-  Common.vcxproj -> C:\Users\Jeremy Yeo\Desktop\Studio Project 2\Debug\Common.lib
+﻿  Common.vcxproj -> D:\Users\539\Documents\GitHub\SP2-Team-18\Debug\Common.lib
