+++ Common/Debug/Common.tlog/Common.lastbuildstate
@@ -1,2 +1,2 @@
 #TargetFrameworkVersion=v4.0:PlatformToolSet=v141:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=10.0.17134.0
-Debug|Win32|C:\Users\Jeremy Yeo\Desktop\Studio Project 2\|
+Debug|Win32|D:\Users\539\Documents\GitHub\SP2-Team-18\|
