+++ DM2122Practical/Debug/DM2122Practical.log
@@ -1,16 +1,4 @@
-﻿  Application.cpp
-  Camera.cpp
-  Camera3.cpp
-  LoadOBJ.cpp
-  LoadTGA.cpp
-  main.cpp
-  Mesh.cpp
-  MeshBuilder.cpp
-  Scene1.cpp
-c:\users\jeremy yeo\desktop\studio project 2\dm2122practical\source\scene1.cpp(93): warning C4552: '-': result of expression not used
-c:\users\jeremy yeo\desktop\studio project 2\dm2122practical\source\scene1.cpp(96): warning C4244: 'argument': conversion from 'Light::LIGHT_TYPE' to 'GLfloat', possible loss of data
-  shader.cpp
-  Utility.cpp
-  Generating Code...
-LINK : warning LNK4098: defaultlib 'MSVCRT' conflicts with use of other libs; use /NODEFAULTLIB:library
-  DM2122Practical.vcxproj -> C:\Users\Jeremy Yeo\Desktop\Studio Project 2\Debug\DM2122Practical.exe
+﻿  SceneGame.cpp
+d:\users\539\documents\github\sp2-team-18\dm2122practical\source\scenegame.cpp(93): warning C4552: '-': result of expression not used
+d:\users\539\documents\github\sp2-team-18\dm2122practical\source\scenegame.cpp(96): warning C4244: 'argument': conversion from 'Light::LIGHT_TYPE' to 'GLfloat', possible loss of data
+  DM2122Practical.vcxproj -> D:\Users\539\Documents\GitHub\SP2-Team-18\Debug\DM2122Practical.exe
