+++ DM2122Practical/Source/Application.cpp
@@ -10,7 +10,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#include "Scene1.h"
+#include "SceneGame.h"
 
 
 GLFWwindow* m_window;
@@ -100,8 +100,9 @@ void Application::Init()
 void Application::Run()
 {
 	//Main Loop
-	Scene *scene = new Scene1();
+	Scene *scene = new SceneGame();
 	scene->Init();
+
 
 	m_timer.startTimer();    // Start timer to calculate how long it takes to render this frame
 	while (!glfwWindowShouldClose(m_window) && !IsKeyPressed(VK_ESCAPE))
